connect_db()

+++ is_like.php
@@ -1,17 +1,11 @@
 <?php
+	include 'connect_db.php';
 	session_start();
 
 	$rid = $_POST["rid"];
 	$is_like = $_POST["like"];
 	
-	$host="mysql17.000webhost.com"; // Host name
-	$username="a6591147_jinghao"; // Mysql username
-	$password="admin123"; // Mysql password
-	$db_name="a6591147_mydata"; // Database name
-
-	// Connect to server and select databse.
-	mysql_connect("$host", "$username", "$password")or die("cannot connect");
-	mysql_select_db("$db_name")or die("cannot select DB");
+	connect_db();
 	
 	// Define $access_token
 	if ( isset($_SESSION['SESS_ACCESS_TOKEN']) ) {
@@ -19,15 +13,17 @@
 	} else {
 		$access_token=$_REQUEST["access_token"]; // for Android
 	}
-	$getid="SELECT id FROM members WHERE access_token='$access_token'";
-	$result = mysql_query($getid);
-	$row = mysql_fetch_assoc($result);
-	$uid = $row['id'];
-	
-	if ($is_like==1) {
-		$sql="INSERT likes VALUES('$uid','$rid')";
-	} elseif ($is_like==0) {
-		$sql="DELETE FROM likes WHERE uid='$uid' AND rid='$rid'";
+	if ( !empty($access_token) ) {
+		$getid="SELECT id FROM members WHERE access_token='$access_token'";
+		$result = mysql_query($getid);
+		$row = mysql_fetch_assoc($result);
+		$uid = $row['id'];
+		
+		if ($is_like==1) {
+			$sql="INSERT likes VALUES('$uid','$rid')";
+		} elseif ($is_like==0) {
+			$sql="DELETE FROM likes WHERE uid='$uid' AND rid='$rid'";
+		}
+		mysql_query($sql);
 	}
-	mysql_query($sql);
 ?>
